requirements.txt
flask==3.0.2
python-dotenv==1.0.1
slack-sdk==3.27.1
requests==2.31.0
typing-extensions==4.9.0
